Update logic of when contacts are deemed equal
Before both the name and the email address had to be equal, but now when
either contact has no name, the contacts will also be deemed equal when
the email addresses are the same.

This allows for looser checks against contacts, where the name may be
supplied in a check, but when it's not only the email address will be
considered. However, when a name is supplied, it must equal the name of
the other contact.